refactor(track-peek): extract audio preview hook
