+++ src/extensions/track-peek/audio-preview.js
@@ -1,17 +1,9 @@
-import { useEffect, useLayoutEffect, useRef, useState } from 'react';
+import { useLayoutEffect, useState } from 'react';
 import { createPortal } from 'react-dom';
-import {
-  ContextMenu,
-  originPlayer,
-  showNotification,
-  SVGIcons,
-  URI,
-} from '@/lib/spicetify.js';
+import { originPlayer, SVGIcons } from '@/lib/spicetify.js';
 import { TooltipWrapper } from '@/lib/host-components.js';
 import { MEDIA_STATUS } from '@/config/constants.js';
-import { getTrack, getAlbumFeed } from '@/services/index.js';
-import { useVolumeSync } from '@/hooks/host/use-volume-sync.js';
-import { useMediaProgress } from '../../hooks/utils/use-media-progress.js';
+import { useAudioPreview } from './use-audio-preview.js';
 import SVGButton from '@/components/shared/svg-button.js';
 import { ProgressCircle } from '@/components/shared/progress/index.js';
 import './audio-preview.css';
@@ -20,64 +12,12 @@ const activeClassName = 'tp-audio-preview-active';
 const inClassName = 'tp-audio-preview-in';
 
 const AudioPreview = ({ container, playStatus }) => {
-  const [status, setStatus] = useState(MEDIA_STATUS.IDLE);
   const [isRendered, setIsRendered] = useState(false);
-  const [currentTrack, setCurrentTrack] = useState(null);
-  const audioRef = useRef(null);
-  const isAudioActiveRef = useRef(false);
-  const snapshotRef = useRef(null);
-  const cacheMapRef = useRef(null);
 
-  const progress = useMediaProgress(audioRef.current);
-
-  useVolumeSync((volume) => {
-    const audio = audioRef.current;
-    if (!audio) return;
-    audio.volume = volume;
+  const { status, track, progress, exit } = useAudioPreview({
+    playStatus,
+    onStart: () => setIsRendered(true),
   });
-
-  const cleanAudio = () => {
-    isAudioActiveRef.current = false;
-    setStatus(MEDIA_STATUS.IDLE);
-    setCurrentTrack(null);
-    const audio = audioRef.current;
-    if (!audio) return;
-    audio.pause();
-    audio.src = '';
-    audio.onended = null;
-  };
-
-  const interruptPlayback = () => {
-    if (snapshotRef.current) return;
-    const mainPlaybackStatus = !originPlayer.getState().isPaused;
-    if (mainPlaybackStatus) {
-      originPlayer.pause();
-    }
-    const playingMedia = new Set();
-    const mediaElements = document.querySelectorAll('video, audio');
-    for (const element of mediaElements) {
-      if (element.paused || element.ended) continue;
-      playingMedia.add(element);
-      element.pause();
-    }
-    snapshotRef.current = {
-      mainPlaybackStatus,
-      playingMedia,
-    };
-  };
-
-  const resumePlayback = () => {
-    const snapshot = snapshotRef.current;
-    if (!snapshot) return;
-    if (snapshot.mainPlaybackStatus) {
-      originPlayer.resume();
-    }
-    for (const element of snapshot.playingMedia) {
-      if (!element.isConnected) continue;
-      element.play().catch(console.warn);
-    }
-    snapshotRef.current = null;
-  };
 
   const handleTransitionEnd = (event) => {
     if (
@@ -103,132 +43,6 @@ const AudioPreview = ({ container, playStatus }) => {
     };
   }, [container, status]);
 
-  useEffect(() => {
-    audioRef.current = new Audio();
-    cacheMapRef.current = new Map();
-
-    const MENU_ITEM_NAME = 'Track peek';
-    const MAX_CACHE_SIZE = 1000;
-    let clickId = 0;
-
-    const onMenuItemClick = async ([trackUri]) => {
-      const currentClickId = ++clickId;
-      interruptPlayback();
-      cleanAudio();
-      isAudioActiveRef.current = true;
-      setStatus(MEDIA_STATUS.LOADING);
-      setIsRendered(true);
-
-      try {
-        if (!cacheMapRef.current.has(trackUri)) {
-          const { album } = await getTrack(trackUri);
-          const { tracks: feedTracks, album: feedAlbum } =
-            await getAlbumFeed(album.uri, album.tracksCount);
-
-          let previewOffset = 0;
-          for (let i = 0; i < album.tracks.length; i++) {
-            const { track } = album.tracks[i];
-            const feedTrack = feedTracks[previewOffset];
-            let name = '';
-            let artists = [];
-            let coverUrl = '';
-            let previewUrl = '';
-            if (
-              previewOffset < feedTracks.length &&
-              feedTrack.uri === track.uri
-            ) {
-              previewOffset++;
-              name = feedTrack.name;
-              artists = feedTrack.artists.map(({ name }) => name);
-              coverUrl = feedAlbum.coverArt[0]?.url;
-              const url = feedTrack.audioPreview.url;
-              if (url) previewUrl = url;
-            }
-            cacheMapRef.current.set(track.uri, {
-              name,
-              artists,
-              coverUrl,
-              previewUrl,
-            });
-          }
-
-          const overflowCount =
-            cacheMapRef.current.size - MAX_CACHE_SIZE;
-          if (overflowCount > 0) {
-            const iterator = cacheMapRef.current.keys();
-            for (let i = 0; i < overflowCount; i++) {
-              cacheMapRef.current.delete(iterator.next().value);
-            }
-          }
-        }
-
-        if (!(isAudioActiveRef.current && clickId === currentClickId))
-          return;
-        const { name, artists, coverUrl, previewUrl } =
-          cacheMapRef.current.get(trackUri) ?? {};
-        if (!previewUrl) {
-          showNotification('[Track Peek]: No audio preview available.');
-          cleanAudio();
-          resumePlayback();
-          return;
-        }
-        setStatus(MEDIA_STATUS.PLAYING);
-        setCurrentTrack({ trackUri, name, artists, coverUrl });
-        audioRef.current.src = previewUrl;
-        audioRef.current.onended = () => {
-          cleanAudio();
-          resumePlayback();
-        };
-        await audioRef.current.play();
-      } catch (error) {
-        if (!(isAudioActiveRef.current && clickId === currentClickId))
-          return;
-        console.error('[Track Peek]:', error);
-        showNotification(
-          '[Track Peek]: Failed to load audio preview.',
-          true
-        );
-        cleanAudio();
-        resumePlayback();
-      }
-    };
-
-    const menuItem = new ContextMenu.Item(
-      MENU_ITEM_NAME,
-      onMenuItemClick,
-      ([uri]) =>
-        URI.isTrack(uri) &&
-        !document.querySelector('[data-testid="watch-feed-view"]'),
-      SVGIcons.nowPlaying
-    );
-    menuItem.register();
-
-    return () => {
-      menuItem.deregister();
-      cleanAudio();
-      resumePlayback();
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!playStatus) return;
-    snapshotRef.current = null;
-    cleanAudio();
-  }, [playStatus]);
-
-  useEffect(() => {
-    const handleMediaPlay = (event) => {
-      if (event.target === audioRef.current) return;
-      snapshotRef.current = null;
-      cleanAudio();
-    };
-
-    document.addEventListener('play', handleMediaPlay, true);
-    return () => {
-      document.removeEventListener('play', handleMediaPlay, true);
-    };
-  }, []);
-
   const disabled = status !== MEDIA_STATUS.PLAYING;
 
   const generalControls = [
@@ -236,9 +50,8 @@ const AudioPreview = ({ container, playStatus }) => {
       icon: 'addToQueue',
       label: 'Add to queue',
       onClick: async () => {
-        await originPlayer.addToQueue([{ uri: currentTrack.trackUri }]);
-        cleanAudio();
-        resumePlayback();
+        await originPlayer.addToQueue([{ uri: track.trackUri }]);
+        exit();
       },
       disabled,
     },
@@ -246,9 +59,7 @@ const AudioPreview = ({ container, playStatus }) => {
       icon: 'play',
       label: 'Play now',
       onClick: async () => {
-        await originPlayer.playAsNextInQueue([
-          { uri: currentTrack.trackUri },
-        ]);
+        await originPlayer.playAsNextInQueue([{ uri: track.trackUri }]);
       },
       disabled,
     },
@@ -265,14 +76,14 @@ const AudioPreview = ({ container, playStatus }) => {
         <TooltipWrapper
           label={
             <>
-              {currentTrack?.name}
-              {currentTrack?.artists && (
+              {track?.name}
+              {track?.artists && (
                 <span
                   style={{
                     fontSize: 12,
                     color: 'var(--spice-subtext)',
                   }}
-                >{` · ${currentTrack.artists.join(', ')}`}</span>
+                >{` · ${track.artists.join(', ')}`}</span>
               )}
             </>
           }
@@ -281,9 +92,9 @@ const AudioPreview = ({ container, playStatus }) => {
         >
           <div className="tp-audio-preview-metadata">
             <div className="tp-audio-preview-cover">
-              {currentTrack?.coverUrl && (
+              {track?.coverUrl && (
                 <img
-                  src={currentTrack.coverUrl}
+                  src={track.coverUrl}
                   alt="cover"
                   width={24}
                   height={24}
@@ -291,7 +102,7 @@ const AudioPreview = ({ container, playStatus }) => {
               )}
             </div>
             <span className="tp-audio-preview-title">
-              {currentTrack?.name ?? 'Loading...'}
+              {track?.name ?? 'Loading...'}
             </span>
           </div>
         </TooltipWrapper>
@@ -312,10 +123,7 @@ const AudioPreview = ({ container, playStatus }) => {
               icon={SVGIcons.pause}
               svgProps={{ width: 12, height: 12 }}
               tooltipProps={{ label: 'Stop' }}
-              onClick={() => {
-                cleanAudio();
-                resumePlayback();
-              }}
+              onClick={exit}
             />
           </ProgressCircle>
         </div>
